fetching data from both APIs works

+++ src/main/java/sunshine/WeatherForecastBackend/model/Weather.java
@@ -18,7 +18,7 @@ public class Weather {
     double temperature;
     double apparentTemperature;
     int clouds;
-    int humidity;
+    double humidity;
     double precipitation;
     double snow;
 }
